feature:insertRoutine 모듈의 View 기반 코드 삭제 (#139)
* #138 - remove : insertRoutine의 insertRoutine 관련 View 삭제

* #138 - remove : 반복 루틴 관련 View 코드 삭제

* #138 - style : 반복 루틴 관련 screen insertRoutine 패키지 하위로 이동
