--- feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/InsertRoutineActionHandler.kt
@@ -1,8 +1,0 @@
-package com.hegunhee.routiner.insertRoutine
-
-interface InsertRoutineActionHandler {
-
-    fun onInsertRoutineButtonClick()
-
-    fun onCategoryAddClick()
-}
