--- feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/InsertRoutineNavigationAction.kt
@@ -1,6 +1,0 @@
-package com.hegunhee.routiner.insertRoutine
-
-sealed interface InsertRoutineNavigationAction {
-
-    object InsertRoutine : InsertRoutineNavigationAction
-}
