+++ feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/common/InsertCommonScreen.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.routiner.insertRoutine.screen.common
+package com.hegunhee.routiner.insertRoutine.common
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
@@ -19,7 +19,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import com.hegunhee.routiner.insertRoutine.R
-import com.hegunhee.routiner.insertRoutine.screen.daily.NormalTextField
+import com.hegunhee.routiner.insertRoutine.daily.NormalTextField
 
 @Composable
 internal fun ColumnScope.RoutineTextEnterScreen(
