+++ feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/daily/InsertRoutineRootScreen.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.routiner.insertRoutine.screen.daily
+package com.hegunhee.routiner.insertRoutine.daily
 
 import android.content.Context
 import android.widget.Toast
@@ -37,10 +37,10 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.repeatOnLifecycle
 import com.hegunhee.routiner.insertRoutine.R
-import com.hegunhee.routiner.insertRoutine.screen.common.CategoryDescriptionScreen
-import com.hegunhee.routiner.insertRoutine.screen.common.CategoryLazyList
-import com.hegunhee.routiner.insertRoutine.screen.common.CategoryTextEnterScreen
-import com.hegunhee.routiner.insertRoutine.screen.common.RoutineTextEnterScreen
+import com.hegunhee.routiner.insertRoutine.common.CategoryDescriptionScreen
+import com.hegunhee.routiner.insertRoutine.common.CategoryLazyList
+import com.hegunhee.routiner.insertRoutine.common.CategoryTextEnterScreen
+import com.hegunhee.routiner.insertRoutine.common.RoutineTextEnterScreen
 import kotlinx.coroutines.flow.SharedFlow
 
 @Composable
