+++ feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/daily/InsertRoutineUiState.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.routiner.insertRoutine.screen.daily
+package com.hegunhee.routiner.insertRoutine.daily
 
 import hegunhee.routiner.model.Category
