+++ feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/navigation/InsertNavigation.kt
@@ -1,10 +1,10 @@
-package com.hegunhee.routiner.insertRoutine.screen.navigation
+package com.hegunhee.routiner.insertRoutine.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.hegunhee.routiner.insertRoutine.screen.daily.InsertRoutineRootScreen
-import com.hegunhee.routiner.insertRoutine.screen.repeat.InsertRepeatRoutineRootScreen
+import com.hegunhee.routiner.insertRoutine.daily.InsertRoutineRootScreen
+import com.hegunhee.routiner.insertRoutine.repeat.InsertRepeatRoutineRootScreen
 
 const val INSERT_ROUTINE_ROUTE = "INSERT_ROUTINE"
