--- feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/repeat/InsertRepeatRoutineActionHandler.kt
@@ -1,7 +1,0 @@
-package com.hegunhee.routiner.insertRoutine.repeat
-
-import com.hegunhee.routiner.insertRoutine.InsertRoutineActionHandler
-
-interface InsertRepeatRoutineActionHandler : InsertRoutineActionHandler{
-
-}
