--- feature/insertRoutine/src/main/java/com/hegunhee/routiner/insertRoutine/repeat/InsertRepeatRoutineNavigationAction.kt
@@ -1,6 +1,0 @@
-package com.hegunhee.routiner.insertRoutine.repeat
-
-sealed interface InsertRepeatRoutineNavigationAction {
-
-    object InsertRepeatRoutine : InsertRepeatRoutineNavigationAction
-}
