--- feature/insertRoutine/src/main/res/layout/fragment_insert_routine.xml
@@ -1,134 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.hegunhee.routiner.insertRoutine.InsertRoutineViewModel" />
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:padding="10dp"
-        tools:targetApi="33">
-
-        <TextView
-            android:id="@+id/routineHeader"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="루틴 입력:"
-            android:textColor="@color/white"
-            android:textSize="20sp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <EditText
-            android:id="@+id/routineQuery"
-            android:layout_width="0dp"
-            android:layout_height="50dp"
-            android:layout_marginTop="10dp"
-            android:paddingStart="10dp"
-            android:background="@drawable/rounded_shape_gray_20dp"
-            android:backgroundTint="@color/white"
-            android:hint="루틴을 입력해주세요"
-            android:inputType="text"
-            android:maxLines="1"
-            android:text="@={viewModel.routineQuery}"
-            android:textColor="@color/gray"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="@id/routineHeader"
-            app:layout_constraintTop_toBottomOf="@id/routineHeader" />
-
-        <TextView
-            android:id="@+id/categoryHeader"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10dp"
-            android:text="카테고리 추가:"
-            android:textColor="@color/white"
-            android:textSize="20sp"
-            app:layout_constraintStart_toStartOf="@id/routineHeader"
-            app:layout_constraintTop_toBottomOf="@id/routineQuery" />
-
-        <EditText
-            android:id="@+id/categoryAddEditText"
-            android:layout_width="0dp"
-            android:layout_height="48dp"
-            android:layout_marginTop="10dp"
-            android:layout_marginEnd="10dp"
-            android:background="@drawable/rounded_shape_gray_20dp"
-            android:backgroundTint="@color/white"
-            android:textColor="@color/gray"
-            android:hint="카테고리를 추가하세요"
-            android:imeOptions="actionDone"
-            android:inputType="text"
-            android:paddingStart="10dp"
-            android:maxLines="1"
-            android:text="@={viewModel.categoryQuery}"
-            app:layout_constraintEnd_toStartOf="@id/categoryAddButton"
-            app:layout_constraintStart_toStartOf="@id/routineHeader"
-            app:layout_constraintTop_toBottomOf="@id/categoryHeader" />
-
-        <ImageButton
-            android:id="@+id/categoryAddButton"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:background="@drawable/ic_add_circle"
-            android:backgroundTint="@color/sky_blue"
-            android:onClick="@{() -> viewModel.onCategoryAddClick()}"
-            app:layout_constraintBottom_toBottomOf="@id/categoryAddEditText"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/categoryAddEditText" />
-
-        <TextView
-            android:id="@+id/categoryAddHeader"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="카테고리 선택:"
-            app:layout_constraintStart_toStartOf="@id/routineQuery"
-            app:layout_constraintTop_toBottomOf="@id/categoryAddEditText"
-            android:layout_marginTop="10dp"
-            android:textSize="20sp"
-            android:textColor="@color/white" />
-
-        <TextView
-            android:id="@+id/categoryDescSecond"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="3dp"
-            android:text="(선택)"
-            android:textSize="13sp"
-            android:textStyle="bold"
-            android:textColor="@color/white"
-            app:layout_constraintBaseline_toBaselineOf="@id/categoryAddHeader"
-            app:layout_constraintBottom_toBottomOf="@id/categoryAddHeader"
-            app:layout_constraintStart_toEndOf="@id/categoryAddHeader" />
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/categoryRecyclerView"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10dp"
-            android:orientation="horizontal"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/categoryAddHeader" />
-
-        <Button
-            android:id="@+id/saveButton"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:onClick="@{() -> viewModel.onInsertRoutineButtonClick()}"
-            android:padding="10dp"
-            android:text="등록하기"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
