+++ feature/main/src/main/java/com/example/main/app/RoutinerApp.kt
@@ -19,9 +19,9 @@ import com.hegunhee.daily.navigation.DAILY_ROUTE
 import com.hegunhee.daily.navigation.dailyNavGraph
 import com.hegunhee.record.navigation.recordNavGraph
 import com.hegunhee.repeat.navigation.repeatRoutineNavGraph
-import com.hegunhee.routiner.insertRoutine.screen.navigation.INSERT_REPEAT_ROUTINE_ROUTE
-import com.hegunhee.routiner.insertRoutine.screen.navigation.INSERT_ROUTINE_ROUTE
-import com.hegunhee.routiner.insertRoutine.screen.navigation.insertNavGraph
+import com.hegunhee.routiner.insertRoutine.navigation.INSERT_REPEAT_ROUTINE_ROUTE
+import com.hegunhee.routiner.insertRoutine.navigation.INSERT_ROUTINE_ROUTE
+import com.hegunhee.routiner.insertRoutine.navigation.insertNavGraph
 import com.hegunhee.setting.navigation.settingNavGraph
 import kotlinx.coroutines.CoroutineScope
