+++ feature/main/src/main/java/com/example/main/app/RoutinerAppScaffoldState.kt
@@ -10,8 +10,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.example.main.drawer.DrawerItem
 import com.example.main.screen.navigation.MAIN_ROUTE
-import com.hegunhee.routiner.insertRoutine.screen.navigation.navigateInsertRepeatRoutine
-import com.hegunhee.routiner.insertRoutine.screen.navigation.navigateInsertRoutine
+import com.hegunhee.routiner.insertRoutine.navigation.navigateInsertRepeatRoutine
+import com.hegunhee.routiner.insertRoutine.navigation.navigateInsertRoutine
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
